+++ app/Http/Controllers/UserController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\UpdateUserRequest;
 use App\Enums\UserJabatan;
 use App\Http\Resources\KPIMResource;
 use App\MyConstant;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -20,6 +21,7 @@ class UserController extends Controller
     */
     public function index()
     {
+        $this->authorize('can-viewAny-user');
         $users = User::all();
 
         return response([
@@ -37,6 +39,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize('can-view-user');
         return response([
             'status' => true,
             'user' => new KPIMResource($user),
@@ -87,6 +90,7 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize('can-update-user');
         $validated = $request->validated();
 
         $user->update($validated);
@@ -105,6 +109,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('can-delete-user');
         $user->delete();
 
         return response([
@@ -121,7 +126,7 @@ class UserController extends Controller
 
         $user = User::where('username', $validated['username']);
         $simpananWajib = $user->simpanan_wajib;
-        
+
         return response([
             'status' => true,
             'simpananWajib' => new KPIMResource($simpananWajib),
@@ -137,7 +142,7 @@ class UserController extends Controller
 
         $user = User::where('username', $validated['username']);
         $simpananPokok = $user->simpanan_pokok;
-        
+
         return response([
             'status' => true,
             'simpananPokok' => new KPIMResource($simpananPokok),
@@ -153,7 +158,7 @@ class UserController extends Controller
 
         $user = User::where('username', $validated['username']);
         $simpananSukarela = $user->simpanan_sukarela;
-        
+
         return response([
             'status' => true,
             'simpananSukarela' => new KPIMResource($simpananSukarela),
